Consolidate API configuration
