--- src/main/kotlin/dev/pankowski/garcon/api/WebConfiguration.kt
@@ -1,13 +1,0 @@
-package dev.pankowski.garcon.api
-
-import org.springframework.stereotype.Component
-import org.springframework.web.method.support.HandlerMethodArgumentResolver
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Component
-class WebConfiguration : WebMvcConfigurer {
-
-  override fun addArgumentResolvers(resolvers: MutableList<HandlerMethodArgumentResolver>) {
-    resolvers.add(SlashCommandMethodArgumentResolver())
-  }
-}
